+++ android/app/src/main/kotlin/tech/mbkm/energrow/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.plts_monitoring
+package tech.mbkm.energrow
 
 import io.flutter.embedding.android.FlutterFragmentActivity